+++ src/spells/sanitize.cpp
@@ -200,12 +200,12 @@ bool spSanitizeBlockOrder::isApplicable( const NifModel *, const QModelIndex & i
 bool spSanitizeBlockOrder::childBeforeParent( NifModel * nif, qint32 block )
 {
 	// get index to the block
-	QModelIndex iBlock( nif->getBlockIndex( block ) );
+	const NifItem * iBlock = nif->getBlockItem( block );
 	// check its type
-	return (
-		nif->blockInherits( iBlock, "bhkRefObject" )
-		&& !nif->blockInherits( iBlock, { "bhkConstraint", "bhkAction" } )
-	);
+	if ( !( iBlock && nif->blockInherits( iBlock, "bhkRefObject" ) ) )
+		return false;
+	return ( !nif->blockInherits( iBlock, { "bhkConstraint", "bhkAction" } )
+			&& !nif->isNiBlock( iBlock, "bhkBallSocketConstraintChain" ) );
 }
 
 // build the nif tree at node block; the block itself and its children are recursively added to
@@ -218,14 +218,23 @@ void spSanitizeBlockOrder::addTree( NifModel * nif, qint32 block, QList<qint32> 
 
 	// special case: add bhkConstraint entities before bhkConstraint
 	// (these are actually links, not refs)
-	QModelIndex iBlock( nif->getBlockIndex( block ) );
-
-	if ( nif->blockInherits( iBlock, "bhkConstraint" ) ) {
-		for ( const auto entity : nif->getLinkArray( iBlock, "Entities" ) ) {
-			addTree( nif, entity, newblocks );
+	const NifItem * iBlock = nif->getBlockItem( block );
+	if ( !iBlock ) {
+		return;
+	} else {
+		QVector<qint32> entities;
+		if ( nif->blockInherits( iBlock, { "bhkConstraint", "bhkBinaryAction" } ) ) {
+			entities.resize( 2 );
+			entities[0] = nif->getLink( iBlock, "Entity A" );
+			entities[1] = nif->getLink( iBlock, "Entity B" );
+		} else if ( nif->blockInherits( iBlock, "bhkUnaryAction" ) ) {
+			entities.append( nif->getLink( iBlock, "Entity" ) );
+		} else if ( nif->isNiBlock( iBlock, "bhkBallSocketConstraintChain" ) ) {
+			entities = nif->getLinkArray( nif->getItem( iBlock, "Constraint Chain Info" ), "Chained Entities" );
 		}
+		for ( const auto entity : entities )
+			addTree( nif, entity, newblocks );
 	}
-
 
 	// add all children of block that should be before block
 	for ( const auto child : nif->getChildLinks( block ) ) {